HADOOP-18181. move prefetch common classes to hadoop common
