+++ hadoop-common-project/hadoop-common/src/test/java/org/apache/hadoop/fs/impl/prefetch/TestBufferPool.java
@@ -17,12 +17,10 @@
  * under the License.
  */
 
-package org.apache.hadoop.fs.common;
+package org.apache.hadoop.fs.impl.prefetch;
 
 import org.junit.Test;
 
-import org.apache.hadoop.fs.s3a.statistics.S3AInputStreamStatistics;
-import org.apache.hadoop.fs.s3a.statistics.impl.EmptyS3AStatisticsContext;
 import org.apache.hadoop.test.AbstractHadoopTestBase;
 
 import static org.junit.Assert.assertEquals;
@@ -34,34 +32,35 @@ public class TestBufferPool extends AbstractHadoopTestBase {
 
   private static final int POOL_SIZE = 2;
   private static final int BUFFER_SIZE = 10;
-  private final S3AInputStreamStatistics s3AInputStreamStatistics =
-      new EmptyS3AStatisticsContext().newInputStreamStatistics();
+
+  private final PrefetchingStatistics statistics =
+      EmptyPrefetchingStatistics.getInstance();
 
   @Test
   public void testArgChecks() throws Exception {
     // Should not throw.
-    BufferPool pool = new BufferPool(POOL_SIZE, BUFFER_SIZE, s3AInputStreamStatistics);
+    BufferPool pool = new BufferPool(POOL_SIZE, BUFFER_SIZE, statistics);
 
     // Verify it throws correctly.
     ExceptionAsserts.assertThrows(
         IllegalArgumentException.class,
         "'size' must be a positive integer",
-        () -> new BufferPool(0, 10, s3AInputStreamStatistics));
+        () -> new BufferPool(0, 10, statistics));
 
     ExceptionAsserts.assertThrows(
         IllegalArgumentException.class,
         "'size' must be a positive integer",
-        () -> new BufferPool(-1, 10, s3AInputStreamStatistics));
+        () -> new BufferPool(-1, 10, statistics));
 
     ExceptionAsserts.assertThrows(
         IllegalArgumentException.class,
         "'bufferSize' must be a positive integer",
-        () -> new BufferPool(10, 0, s3AInputStreamStatistics));
+        () -> new BufferPool(10, 0, statistics));
 
     ExceptionAsserts.assertThrows(
         IllegalArgumentException.class,
         "'bufferSize' must be a positive integer",
-        () -> new BufferPool(1, -10, s3AInputStreamStatistics));
+        () -> new BufferPool(1, -10, statistics));
 
     ExceptionAsserts.assertThrows(
         NullPointerException.class,
@@ -85,7 +84,7 @@ public class TestBufferPool extends AbstractHadoopTestBase {
 
   @Test
   public void testGetAndRelease() {
-    BufferPool pool = new BufferPool(POOL_SIZE, BUFFER_SIZE, s3AInputStreamStatistics);
+    BufferPool pool = new BufferPool(POOL_SIZE, BUFFER_SIZE, statistics);
     assertInitialState(pool, POOL_SIZE);
 
     int count = 0;
@@ -132,7 +131,7 @@ public class TestBufferPool extends AbstractHadoopTestBase {
   private void testReleaseHelper(BufferData.State stateBeforeRelease, boolean expectThrow)
       throws Exception {
 
-    BufferPool pool = new BufferPool(POOL_SIZE, BUFFER_SIZE, s3AInputStreamStatistics);
+    BufferPool pool = new BufferPool(POOL_SIZE, BUFFER_SIZE, statistics);
     assertInitialState(pool, POOL_SIZE);
 
     BufferData data = this.acquire(pool, 1);
